refactor(clickhouse): fold listBranches into tracing.ts
branches are a subset of traces; the dedicated branches.ts module was
gratuitous separation. Move the listBranches function (with its scope
filter, dedupeKey tie-breaker, and BRANCH_SPAN_TYPES guard) into
tracing.ts next to the other read paths over trace data, drop the
branches.ts file, and update v-next/index.ts to call
tracingOps.listBranches.

Note: trace-roots.ts (which owns listTraces / getRootSpan) is left as a
separate module since it's pre-existing and out of scope for this PR.
